requirements.txt
aiogram==3.3.0
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
peewee==3.16.3
phonenumbers==8.13.22
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
redis==5.0.1
typing_extensions==4.7.1
yarl==1.9.2
pytest==7.4.3
faker==22.2.0
aiocron==1.8

